fix: add static params for book pages
