+++ app/book/[id]/EditorPage.tsx
@@ -1,0 +1,197 @@
+'use client';
+
+import { useEffect, useState } from 'react';
+import { Button } from '@openameba/spindle-ui';
+import { Menu, X } from 'lucide-react';
+import { useStore } from '@/lib/store';
+import Header from '@/components/Header';
+import Panels from '@/components/Panels';
+import SourceManager from '@/components/Left/SourceManager';
+import SourceChat from '@/components/Left/SourceChat';
+import TitleBar from '@/components/Editor/TitleBar';
+import RichEditor from '@/components/Editor/RichEditor';
+import DictionarySearch from '@/components/Right/DictionarySearch';
+import MaterialUpload from '@/components/Right/MaterialUpload';
+import MaterialChat from '@/components/Right/MaterialChat';
+
+export default function EditorPage({ bookId }: { bookId: string }) {
+  
+  const { 
+    books, 
+    ui, 
+    setLeftTab, 
+    setRightTab, 
+    setRightSubTab,
+    setRightPanelOpen,
+    setCurrentBookId
+  } = useStore();
+
+  const [isTablet, setIsTablet] = useState(false);
+
+  const book = books.find(b => b.id === bookId);
+
+  useEffect(() => {
+    setCurrentBookId(bookId);
+    
+    const checkScreenSize = () => {
+      const width = window.innerWidth;
+      setIsTablet(width >= 768 && width < 1280);
+    };
+
+    checkScreenSize();
+    window.addEventListener('resize', checkScreenSize);
+    return () => window.removeEventListener('resize', checkScreenSize);
+  }, [bookId, setCurrentBookId]);
+
+  if (!book) {
+    return (
+      <div className="h-screen flex flex-col bg-gray-50">
+        <Header />
+        <div className="flex-1 flex items-center justify-center">
+          <div className="text-center">
+            <h2 className="text-xl font-semibold text-gray-900 mb-2">
+              ブックが見つかりません
+            </h2>
+            <p className="text-gray-600">
+              指定されたブックは存在しないか、削除されている可能性があります。
+            </p>
+          </div>
+        </div>
+      </div>
+    );
+  }
+
+  // 左パネル
+  const leftPanel = (
+    <div className="h-full flex flex-col">
+      {/* タブ */}
+      <div className="tab-list">
+        <button 
+          className={`tab-button ${ui.leftTab === 'manage' ? 'active' : ''}`}
+          onClick={() => setLeftTab('manage')}
+        >
+          ソース管理
+        </button>
+        <button 
+          className={`tab-button ${ui.leftTab === 'chat' ? 'active' : ''}`}
+          onClick={() => setLeftTab('chat')}
+        >
+          ソースチャット
+        </button>
+      </div>
+
+      {/* コンテンツ */}
+      <div className="flex-1 overflow-hidden">
+        {ui.leftTab === 'manage' ? (
+          <SourceManager bookId={bookId} />
+        ) : (
+          <SourceChat bookId={bookId} />
+        )}
+      </div>
+    </div>
+  );
+
+  // 中央パネル
+  const centerPanel = (
+    <div className="h-full flex flex-col">
+      <TitleBar bookId={bookId} />
+      <div className="flex-1 overflow-hidden">
+        <RichEditor bookId={bookId} />
+      </div>
+    </div>
+  );
+
+  // 右パネル
+  const rightPanel = (
+    <div className="h-full flex flex-col">
+      {/* タブレット用ヘッダー */}
+      {isTablet && (
+        <div className="p-3 border-b border-gray-200 flex justify-between items-center">
+          <h2 className="font-semibold">辞書・資料</h2>
+          <Button
+            variant="lighted"
+            size="small"
+            onClick={() => setRightPanelOpen(false)}
+          >
+            <X className="w-4 h-4" />
+          </Button>
+        </div>
+      )}
+
+      {/* 上位タブ */}
+      <div className="tab-list">
+        <button 
+          className={`tab-button ${ui.rightTab === 'dict' ? 'active' : ''}`}
+          onClick={() => setRightTab('dict')}
+        >
+          辞書・表現検索
+        </button>
+        <button 
+          className={`tab-button ${ui.rightTab === 'material' ? 'active' : ''}`}
+          onClick={() => setRightTab('material')}
+        >
+          資料検索
+        </button>
+      </div>
+
+      {/* コンテンツ */}
+      <div className="flex-1 overflow-hidden">
+        {ui.rightTab === 'dict' ? (
+          <DictionarySearch bookId={bookId} />
+        ) : (
+          <div className="h-full flex flex-col">
+            {/* ミニタブ */}
+            <div className="tab-list border-b-0">
+              <button 
+                className={`tab-button ${ui.rightSubTab === 'upload' ? 'active' : ''}`}
+                onClick={() => setRightSubTab('upload')}
+              >
+                資料アップロード
+              </button>
+              <button 
+                className={`tab-button ${ui.rightSubTab === 'chat' ? 'active' : ''}`}
+                onClick={() => setRightSubTab('chat')}
+              >
+                資料チャット
+              </button>
+            </div>
+            
+            <div className="flex-1 overflow-hidden">
+              {ui.rightSubTab === 'upload' ? (
+                <MaterialUpload bookId={bookId} />
+              ) : (
+                <MaterialChat bookId={bookId} />
+              )}
+            </div>
+          </div>
+        )}
+      </div>
+    </div>
+  );
+
+  return (
+    <div className="h-screen flex flex-col bg-white">
+      <div className="flex items-center justify-between">
+        <Header />
+        {/* タブレット用右パネル開閉ボタン */}
+        {isTablet && (
+          <Button
+            variant="lighted"
+            size="small"
+            onClick={() => setRightPanelOpen(!ui.rightPanelOpen)}
+          >
+            <Menu className="w-5 h-5" />
+          </Button>
+        )}
+      </div>
+      
+      <div className="flex-1 overflow-hidden">
+        <Panels
+          leftPanel={leftPanel}
+          centerPanel={centerPanel}
+          rightPanel={rightPanel}
+        />
+      </div>
+    </div>
+  );
+}
