+++ app/layout.tsx
@@ -1,6 +1,7 @@
-import '@openameba/spindle-ui/index.css';
 import './globals.css';
 import type { Metadata } from 'next';
+import ThemeRegistry from '@/components/ThemeRegistry';
+import Header from '@/components/Header';
 
 export const metadata: Metadata = {
   title: 'My Books Editor',
@@ -14,7 +15,12 @@ export default function RootLayout({
 }) {
   return (
     <html lang="ja">
-      <body className="font-sans">{children}</body>
+      <body className="font-sans">
+        <ThemeRegistry>
+          <Header />
+          {children}
+        </ThemeRegistry>
+      </body>
     </html>
   );
 }
