+++ components/ThemeRegistry.tsx
@@ -1,0 +1,18 @@
+'use client';
+
+import { ReactNode } from 'react';
+import { ThemeProvider, createTheme } from '@mui/material/styles';
+import CssBaseline from '@mui/material/CssBaseline';
+
+const theme = createTheme({
+  // カスタムテーマ設定（必要に応じて）
+});
+
+export default function ThemeRegistry({ children }: { children: ReactNode }) {
+  return (
+    <ThemeProvider theme={theme}>
+      <CssBaseline />
+      {children}
+    </ThemeProvider>
+  );
+}
