+++ package.json
@@ -42,6 +42,10 @@
     "@radix-ui/react-tooltip": "^1.1.2",
     "@tiptap/react": "^3.2.1",
     "@tiptap/starter-kit": "^3.2.1",
+    "@mui/material": "^5.15.15",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.15",
     "@types/node": "20.6.2",
     "@types/react": "18.2.22",
     "@types/react-dom": "18.2.7",
